requirements.txt
ipython==9.0.2
Pillow==11.1.0
pydantic==2.11.2
pydantic_ai==0.0.52
pydantic_graph==0.0.52
python-dotenv==1.1.0
Requests==2.32.3
Spire.Doc.Free==12.12.0
tavily_python==0.5.1
tabulate==0.9.0
